requirements.txt
absl-py==1.4.0
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
colored==1.4.4
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
exceptiongroup==1.1.1
executing==1.2.0
fg==0.0.4
flatbuffers==23.3.3
fonttools==4.39.3
frozendict==2.3.6
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
html5lib==1.1
idna==3.4
imageio==2.27.0
imgaug==0.4.0
importlib-metadata==6.3.0
instagrapi==1.17.3
ipykernel==6.22.0
ipython==8.12.0
jax==0.4.8
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kaggle==1.5.13
keras==2.12.0
Keras-Applications==1.0.8
keras-ocr==0.8.9
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
libs==0.0.10
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
multitasking==0.0.11
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==2.0.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.4.0
pingo==0.2.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.7
pyee==8.2.2
Pygments==2.14.0
pyparsing==3.0.9
pyppeteer==1.0.2
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.8.3
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.6.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tesseract==0.1.3
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.3.21
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.3
wrapt==1.14.1
wsproto==1.2.0
yfinance==0.2.14
zipp==3.15.0